moved cuda kernel
