+++ inference_lib/src/aqlm/__init__.py
@@ -1,3 +1,2 @@
-import aqlm.cuda
 import aqlm.inference_kernels
 from aqlm.inference import QuantizedLinear
